Add ikuasAPI Login
We need to get some data from KUAS AP Service. That' why we need to
login ikuasAPI, we can get data conveniently.

We used ikuass_API modules which are in modules folder. It's a class
which connect API easily. And it pass data with encryption key. It's a
little safer than passing password without encryption.